+++ NEWS.md
@@ -1,5 +1,8 @@
 # nanonext 0.1.0.9000 (under development)
 
+* Add `ncurl()` minimalistic http(s) client.
+* Allow setting the environment variable 'NANONEXT_TLS=1' prior to package installation to enable TLS where the system NNG library has been built with TLS support (using Mbed TLS).
+
 # nanonext 0.1.0
 
 * Initial release to CRAN, rOpenSci R-universe and Github.